requirements.txt
setuptools==56.2.0
wheel==0.36.2
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dataclasses-json==0.5.2
html5lib==1.1
htmlmin==0.1.12
idna==2.10
Jinja2==3.0.1
json-schema-for-humans==0.31.0
jsonschema==3.2.0
markdown2==2.4.0
MarkupSafe==2.0.1
marshmallow==3.11.1
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
Pygments==2.8.1
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
stringcase==1.2.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.4
webencodings==0.5.1
rdflib==6.0.2
Pillow==8.4.0
strawberry-graphql==0.151.2